[server] WorkspaceService.startWorkspace - part I: API-side, no tests
